+++ crates/swc_typescript/src/fast_dts/util/mod.rs
@@ -1,0 +1,3 @@
+pub mod ast_ext;
+pub mod expando_function_collector;
+pub mod types;
